Practica Integradora 3 v1

+++ practicaIntegradora3/src/services/products.services.js
@@ -2,6 +2,7 @@ import { productResponseDto } from "../dto/product-response.dto.js";
 import productsRepository from "../persistences/mongo/repositories/products.repository.js";
 import error from "../errors/customErrors.js"
 
+
 // lógica de petición de todos los productos o por un límite
 const getProducts = async (query, options) => {
 	const products = await productsRepository.getProducts(query, options);
@@ -17,8 +18,13 @@ const getProductById = async (id) => {
 };
 
 // lógica de carga de un nuevo producto
-const createProduct = async (data) => {
-	const product = await productsRepository.createProduct(data);
+const createProduct = async (data, user) => {
+    let productData = data;
+    if (user.role === "premium") {
+        productData = { ...data, owner: user._id};
+    };
+
+	const product = await productsRepository.createProduct(productData);
     if (!product) throw error.badRequestError();
     return product;
 };
@@ -31,7 +37,12 @@ const updateProductById = async (id, data) => {
 };
 
 // lógica para eliminar un producto
-const deleteProductById = async (id) => {
+const deleteProductById = async (id, user) => {
+    const productData = await productsRepository.getProductById(id);
+    if (user.role === "premium" && productData.owner !== user._id) {
+        throw error.unauthorizedError("User not authorized to delete product")
+    }
+    
     const product = await productsRepository.deleteProductById(id);
     if (!product) throw error.notFoundError(`Product id:${id} not found`);
